WIP refactor entities
